+++ tests/generator/lib.php
@@ -55,6 +55,10 @@ class quiz_archiver_generator extends \testing_data_generator {
             'backupfile' => '/../fixtures/referencequiz.mbz',
             'quizname' => 'Reference Quiz (standard question types)',
         ],
+        'multiattempt' => [
+            'backupfile' => '/../fixtures/referencequiz-multiattempt.mbz',
+            'quizname' => 'Reference Quiz (standard question types)',
+        ],
         'qtype_jack' => [
             'backupfile' => '/../fixtures/referencequiz-qtype_jack.mbz',
             'quizname' => 'qType JACK Reference Quiz',
@@ -88,7 +92,6 @@ class quiz_archiver_generator extends \testing_data_generator {
             ],
             'settings' => [
                 'num_attempts' => 3,
-                'export_attempts' => 1,
                 'export_report_section_header' => 1,
                 'export_report_section_quiz_feedback' => 1,
                 'export_report_section_question' => 1,
@@ -243,15 +246,10 @@ class quiz_archiver_generator extends \testing_data_generator {
             }
         }
         $quiz = $DB->get_record('quiz', ['id' => $cm->instance], '*', MUST_EXIST);
-        $attemptids = array_values(array_map(
-            fn($r): int => $r->id,
-            $DB->get_records('quiz_attempts', ['quiz' => $quiz->id], '', 'id')
-        ));
+        $attempts = $DB->get_records('quiz_attempts', ['quiz' => $quiz->id], '', 'id, userid');
 
-        $userids = array_values(array_map(
-            fn($r): int => $r->userid,
-            $DB->get_records('quiz_attempts', ['quiz' => $quiz->id], '', 'userid')
-        ));
+        $attemptids = array_values(array_map(fn($r): int => $r->id, $attempts));
+        $userids = array_values(array_map(fn($r): int => $r->userid, $attempts));
 
         return (object) [
             'course' => $course,